% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree_coef.R
\name{agree_coef}
\alias{agree_coef}
\title{Agreeement Coefficents}
\usage{
agree_coef(
  wide = TRUE,
  col.names = NULL,
  measure,
  item,
  id,
  data,
  weighted = FALSE,
  conf.level = 0.95
)
}
\arguments{
\item{wide}{Logical value (TRUE or FALSE) indicating if data is in a "wide" format. Default is TRUE.}

\item{col.names}{If wide is equal to TRUE then col.names is a list of the column names containing the measurements for reliability analysis.}

\item{measure}{Name of column containing the measurement of interest.}

\item{item}{Name of column containing the items. If this is an inter-rater reliability study then this would indicate the rater (e.g., rater1, rater2, rater3, etc).}

\item{id}{Column with subject identifier.}

\item{data}{Data frame with all data.}

\item{conf.level}{the confidence level required. Default is 95\%.}
}
\value{
Returns single data frame of inter-rater reliability coefficients.
}
\description{
agree_coef produces inter-rater reliability or "agreement coefficients" as described by Gwet.
}
\section{References}{

Gwet, K.L. (2014, ISBN:978-0970806284). “Handbook of Inter-Rater Reliability,” 4th Edition. Advanced Analytics, LLC.
}

\examples{
#TBA

}
