% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree_np.R
\name{agree_np}
\alias{agree_np}
\title{Nonparametric Test for Limits of Agreement}
\usage{
agree_np(
  x,
  y,
  id = NULL,
  data,
  delta = NULL,
  prop_bias = FALSE,
  agree.level = 0.95,
  conf.level = 0.95
)
}
\arguments{
\item{x}{Name of column with first measurement.}

\item{y}{Name of other column with first measurement.}

\item{id}{Column with subject identifier with samples are taken in replicates.}

\item{data}{Data frame with all data.}

\item{delta}{The threshold below which methods agree/can be considered equivalent and this argument is required. Equivalence Bound for Agreement or Maximal Allowable Difference.}

\item{prop_bias}{Logical indicator (TRUE/FALSE) of whether proportional bias should be considered for the limits of agreement calculations.}

\item{agree.level}{the agreement level required. Default is 95\%. The proportion of data that should lie between the thresholds, for 95\% limits of agreement this should be 0.95.}

\item{conf.level}{the confidence level required. Default is 95\%.}
}
\value{
Returns adv-agree object with the results of the agreement analysis.

\describe{
  \item{\code{"loa"}}{a data frame of the limits of agreement.}
  \item{\code{"agee"}}{a data frame of the binomial proportion of results in agreement..}
  \item{\code{"h0_test"}}{Decision from hypothesis test.}
  \item{\code{"call"}}{the matched call}

}
}
\description{
agree_np produces an agreement analysis based on quantiles and
}
\section{References}{

TBA
}

\examples{
# TBA
}
