% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreetest.h.R
\name{agreetest}
\alias{agreetest}
\title{Simple Test of Agreement in jamovi}
\usage{
agreetest(
  data,
  method1,
  method2,
  ciWidth = 0.95,
  agreeWidth = 0.95,
  testValue = 2,
  CCC = TRUE
)
}
\arguments{
\item{data}{.}

\item{method1}{.}

\item{method2}{.}

\item{ciWidth}{a number between .50 and .999 (default: .95), the width of
confidence intervals}

\item{agreeWidth}{a number between .50 and .999 (default: .95), the width
of agreement limits}

\item{testValue}{a number specifying the limit of agreement}

\item{CCC}{\code{TRUE} or \code{FALSE} (default), produce CCC table}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$blandtab} \tab \tab \tab \tab \tab a table \cr
  \code{results$ccctab} \tab \tab \tab \tab \tab a table \cr
  \code{results$cites} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$blandtab$asDF}

\code{as.data.frame(results$blandtab)}
}
\description{
Simple Test of Agreement in jamovi
}
