% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandPowerCurve.R
\name{blandPowerCurve}
\alias{blandPowerCurve}
\title{Estimate power curve for Bland-Altman limits of agreement test}
\usage{
blandPowerCurve(
  samplesizes = seq(10, 100, 1),
  mu = 0,
  SD,
  delta,
  conf.level = 0.95,
  agree.level = 0.95
)
}
\arguments{
\item{samplesizes}{vector of samples sizes to}

\item{mu}{mean of differences}

\item{SD}{standard deviation of differences}

\item{delta}{pre-determined clinical agreement interval}

\item{conf.level}{the confidence level required. Default is 95\%.}

\item{agree.level}{the agreement level required. Default is 95\%. The proportion of data that should lie between the thresholds, for 95\% limits of agreement this should be 0.95.}
}
\value{
a dataframe containing the power analysis results
}
\description{
Estimate power curve for Bland-Altman limits of agreement test
}
\examples{
\dontrun{
powerCurve <- blandPowerCurve(samplesizes = seq(10, 200, 1),
mu = 0,
SD = 3.3,
delta = 8,
conf.level = .95,
agree.level = .95)
# Plot the power curve
plot(powerCurve, type = 1)
# Find at what N power of .8 is achieved
find_n(powerCurve, power = .8)
}
}
