% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reps}
\alias{reps}
\title{reps}
\format{
A data frame with 20 rows with 3 variables
\describe{
  \item{id}{Subject identifier}
  \item{x}{X measurement}
  \item{y}{Y measurement}


}
}
\usage{
reps
}
\description{
A fake data set of a agreement study where both measures have replicates.
}
\keyword{datasets}
