% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmvagreemulti.h.R
\name{jmvagreemulti}
\alias{jmvagreemulti}
\title{Agreement Analysis for Nested or Replicate Data}
\usage{
jmvagreemulti(
  data,
  method1,
  method2,
  id,
  ciWidth = 0.95,
  agreeWidth = 0.95,
  testValue = 2,
  valEq = FALSE
)
}
\arguments{
\item{data}{.}

\item{method1}{.}

\item{method2}{.}

\item{id}{.}

\item{ciWidth}{a number between .50 and .999 (default: .95), the width of
confidence intervals}

\item{agreeWidth}{a number between .50 and .999 (default: .95), the width
of agreement limits}

\item{testValue}{a number specifying the limit of agreement}

\item{valEq}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Agreement Analysis for Nested or Replicate Data
}
