% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.loa_mixed_bs.R
\name{loa_mixed_bs-methods}
\alias{loa_mixed_bs-methods}
\alias{print.loa_mixed_bs}
\alias{plot.loa_mixed_bs}
\title{Methods for loa_mixed_bs objects}
\usage{
\method{print}{loa_mixed_bs}(x, ...)

\method{plot}{loa_mixed_bs}(x, ...)
}
\arguments{
\item{x}{object of class \code{loa_mixed_bs} as returned from \code{loa_mixed}}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{loa_mixed}}.}
}
\value{
\describe{
\item{\code{print}}{Prints short summary of the Limits of Agreement}
\item{\code{plot}}{Returns a plot of the limits of agreement}
}
}
\description{
Methods defined for objects returned from the loa_mixed functions.
}
