% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.powerCurve.R
\name{powerCurve-methods}
\alias{powerCurve-methods}
\alias{find_n}
\alias{plot.powerCurve}
\title{Methods for powerCurve objects}
\usage{
find_n(x, power = 0.8, tolerance = 0.01)

\method{plot}{powerCurve}(x, type = 1, ...)
}
\arguments{
\item{x}{object of class \code{powerCurve}}

\item{power}{Level of power (value between 0 and 1) for find_n to find the sample size.}

\item{tolerance}{Degree of tolerance for find_n to find the nearest observed power to the desired power.}

\item{type}{Type of plot to output. Default (1) is power curve plot while type=2 will produce a confidence interval width plot.}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{blandPowerCurve}}.}
}
\value{
\describe{
  \item{\code{plot}}{Returns a plot of the limits of agreement (type = 1) or concordance plot (type = 2)}
  \item{\code{find_n}}{Find sample size at which desired power is achieved}
}
}
\description{
Methods defined for objects returned from the powerCurve function.
}
