% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_agree.R
\name{pwr_agree}
\alias{pwr_agree}
\title{Oower for Absolute Agreement}
\usage{
pwr_agree(
  prop0 = 0.8,
  delta = 0.1,
  power = 0.8,
  mu,
  sigma,
  alpha = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{prop0}{Null central proportion: the propotion of data that should lie between the tresholds, for 95\% limits of agreement this should be 0.95.}

\item{delta}{The treshold below which methods agree/can be considered equivalent, can be in any units. Equivalence Bound for Agreement.}

\item{power}{Nominal power, the probability of rejecting the null-hypothesis when the alternative hypothesis, the specified null proportion of the data lies between the thresholds, is true.}

\item{mu}{Mean difference between methods, based on previous research or pilot work.}

\item{sigma}{SD of difference, based on previous research or pilot work. This SD can be determind from the limits of agreement.}

\item{alpha}{Desired Type 1 error rate. Probability of concluding the methods agree, the specified proportion of data falls between the tresholds, while they actually not agree. Default is .05}

\item{verbose}{Option to print a summary of results to the console.}
}
\value{
Returns single list with the results of the power analysis for agreement.

\describe{
  \item{\code{"gam"}}{Gamma statistic.}
  \item{\code{"power"}}{Estimated power (1-Beta).}
  \item{\code{"alpha"}}{Desired Type 1 error rate.}
  \item{\code{"n"}}{Sample Size Requirement.}
  \item{\code{"mu"}}{Specified mean difference between methods.}
  \item{\code{"sigma"}}{Specified SD of the difference.}
  \item{\code{"delta"}}{Proposed threshold of agreement.}

}
}
\description{
Oower for Absolute Agreement
}
\section{References}{

Gwowen Shieh (2019): Assessing Agreement Between Two Methods of Quantitative Measurements: Exact Test Procedure and Sample Size Calculation, Statistics in Biopharmaceutical Research, <https://doi.org/10.1080/19466315.2019.1677495>
}

\examples{
#to be added

}
