% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.simple_agree.R
\name{simple_agree-methods}
\alias{simple_agree-methods}
\alias{print.simple_agree}
\alias{plot.simple_agree}
\title{Methods for simple_agree objects}
\usage{
\method{print}{simple_agree}(x, ...)

\method{plot}{simple_agree}(x, type = 1)
}
\arguments{
\item{x}{object of class \code{simple_agree} as returned from \code{loa_mixed}}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{loa_mixed}}.}
}
\value{
\describe{
  \item{\code{print}}{Prints short summary of the Limits of Agreement}
  \item{\code{plot}}{Returns a plot of the limits of agreement (type = 1) or concordance plot (type = 2)}
}
}
\description{
Methods defined for objects returned from the loa_mixed functions.
}
