% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.simple_deming.R
\name{simple_deming-methods}
\alias{simple_deming-methods}
\alias{print.simple_deming}
\alias{plot.simple_deming}
\title{Methods for simple_deming objects}
\usage{
\method{print}{simple_deming}(x, ...)

\method{plot}{simple_deming}(x, ...)
}
\arguments{
\item{x}{object of class \code{simple_deming} as returned from a function starting with 'agree'}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{agree_test}}.}
}
\value{
\describe{
  \item{\code{print}}{Prints short summary of the Limits of Agreement}
  \item{\code{plot}}{Returns a plot of the limits of agreement (type = 1) or concordance plot (type = 2)}
  \item{\code{check}}{Returns 2 plots, p_norm and p_het, testing the assumptions of a Bland-Altman analysis. P-values for the normality and heteroskedascity tests are provided as captions to the plot.}
}
}
\description{
Methods defined for objects returned from the agree functions.
}
