% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.simple_reli.R
\name{simple_reli-methods}
\alias{simple_reli-methods}
\alias{print.simple_reli}
\alias{plot.simple_reli}
\alias{check.simple_reli}
\title{Methods for simple_reli objects}
\usage{
\method{print}{simple_reli}(x, ...)

\method{plot}{simple_reli}(x, ...)

\method{check}{simple_reli}(x)
}
\arguments{
\item{x}{object of class \code{simple_reli} as returned from the reli_stats function}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{reli_stats}}.}
}
\value{
\describe{
\item{\code{print}}{Prints short summary of the Limits of Agreement}
\item{\code{plot}}{Returns a plot of the data points used in the reliability analysis}
}
}
\description{
Methods defined for objects returned from the agree functions.
}
