% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimplyAgree-package.R
\docType{package}
\name{SimplyAgree-package}
\alias{SimplyAgree}
\alias{SimplyAgree-package}
\title{SimplyAgree: Flexible and Robust Agreement and Reliability Analyses}
\description{
Reliability and agreement analyses often have limited software support. Therefore, this package was created to make agreement and reliability analyses easier for the average researcher. The functions within this package include simple tests of agreement, agreement analysis for nested and replicate data, and provide robust analyses of reliability. In addition, this package contains a set of functions to help when planning studies looking to assess measurement agreement.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aaroncaldwell.us/SimplyAgree/}
  \item Report bugs at \url{https://github.com/arcaldwell49/SimplyAgree/issues}
}

}
\author{
\strong{Maintainer}: Aaron Caldwell \email{arcaldwell49@gmail.com}

}
\keyword{internal}
