% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance_delta_gls.R
\name{tolerance_delta_gls}
\alias{tolerance_delta_gls}
\title{Calculate the Tolerance Limits from a Agreement Study}
\usage{
tolerance_delta_gls(
  data,
  x,
  y,
  id = NULL,
  condition = NULL,
  time = NULL,
  pred_level = 0.95,
  tol_level = 0.95,
  prop_bias = FALSE,
  log_tf = FALSE,
  cor_type = c("sym", "car1", "ar1", "none"),
  correlation = NULL,
  weights = NULL,
  keep_model = TRUE
)
}
\arguments{
\item{data}{Data frame with all data.}

\item{x}{Name of column with first measurement}

\item{y}{Name of other column with the other measurement to compare to the first.}

\item{id}{Column with subject identifier. Default is "id" if no entry is provided.}

\item{condition}{Column name indicating different conditions subjects were tested under. This can be left missing if there are no differing conditions to be tested.}

\item{time}{A column naming/numbering the time point. Only necessary if the data is from time series collection.}

\item{pred_level}{The prediction level. Default is 95\%.}

\item{tol_level}{The tolerance level related to the prediction level. Default is 95\%.}

\item{prop_bias}{Logical indicator (TRUE/FALSE) of whether proportional bias should be considered for the prediction/tolerance limits calculations.}

\item{log_tf}{Calculate limits of agreement using log-transformed data.}

\item{cor_type}{The type of correlation structure. "sym" is for Compound Symmetry, "car1" is for continuous autocorrelation structure of order 1, or "ar1" for autocorrelation structure of order 1.}

\item{correlation}{an optional corStruct object describing the within-group correlation structure that overrides the default setting. See the documentation of corClasses for a description of the available corStruct classes. If a grouping variable is to be used, it must be specified in the form argument to the corStruct constructor. Defaults to NULL}

\item{weights}{an optional varFunc object or one-sided formula describing the within-group heteroscedasticity structure that overrides the default setting. If given as a formula, it is used as the argument to varFixed, corresponding to fixed variance weights. See the documentation on varClasses for a description of the available varFunc classes.}

\item{keep_model}{Logical indicator to retain the GLS model. Useful when working with large datasets and the model is very large.}
}
\value{
Returns single tolerance_delta class object with the results of the agreement analysis with prediction/tolerance limits..

\describe{
  \item{\code{"limits"}}{A data frame containing the prediction/tolerance limits.}
  \item{\code{"model"}}{The GLS model; NULL if keep_model set to FALSE.}
  \item{\code{"call"}}{The matched call.}
}
}
\description{
A function for calculating tolerance limits for the difference between two measurements. Approximately the same procedure as Bland-Altman limits of agreement.
}
\section{References}{

}

\examples{
data('reps')

}
