% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agree_reps.R
\name{agree_reps}
\alias{agree_reps}
\title{Tests for Absolute Agreement}
\usage{
agree_reps(x, y, id, data, delta, agree.level = 0.95, conf.level = 0.95)
}
\arguments{
\item{x}{Name of column with first measurement}

\item{y}{Name of other column with first measurement}

\item{id}{Column with subject identifier}

\item{data}{Data frame with all data}

\item{delta}{The threshold below which methods agree/can be considered equivalent, can be in any units. Equivalence Bound for Agreement.}

\item{agree.level}{the agreement level required. Default is 95\%. The proportion of data that should lie between the thresholds, for 95\% limits of agreement this should be 0.95.}

\item{conf.level}{the confidence level required. Default is 95\%.}
}
\value{
Returns single list with the results of the agreement analysis.

\describe{
  \item{\code{"loa"}}{a data frame of the limits of agreement including the average difference between the two sets of measurements, the standard deviation of the difference between the two sets of measurements and the lower and upper confidence limits of the difference between the two sets of measurements.}
  \item{\code{"identity.plot"}}{Plot of x and y with a line of identity with a linear regression line}
  \item{\code{"bland_alt.plot"}}{Simple Bland-Altman plot. Red line are the upper and lower bounds for shieh test; grey box is the acceptable limits (delta). If the red lines are within the grey box then the shieh test should indicate 'reject h0', or to reject the null hypothesis that this not acceptable agreement between x & y.}

}
}
\description{
Tests for Absolute Agreement
}
\section{References}{

Zou, G. Y. (2013). Confidence interval estimation for the Bland–Altman limits of agreement with multiple observations per individual. Statistical methods in medical research, 22(6), 630-642.
}

\examples{
#to be added

}
