% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits_of_agree.R
\name{agreement_limit}
\alias{agreement_limit}
\title{Calculate the Limits of Agreement}
\usage{
agreement_limit(
  x,
  y,
  id = NULL,
  data,
  data_type = c("simple", "nest", "reps"),
  loa_calc = c("mover", "blandaltman"),
  agree.level = 0.95,
  alpha = 0.05,
  prop_bias = FALSE,
  log = FALSE
)
}
\arguments{
\item{x}{Name of column with first measurement}

\item{y}{Name of other column with the other measurement to compare to the first.}

\item{id}{Column with subject identifier. Default is "id" if no entry is provided.}

\item{data}{Data frame with all data.}

\item{data_type}{The type of data structure. Options include "simple" (all independent data points), "nest" (nested data) and "reps" (replicated data points).}

\item{loa_calc}{The mehtod by which the limits of agreement confidence intervals are calculated. Options are "mover" (Methods of Recovering Variances method) or "blandlatman" (Bland-Altman method).}

\item{agree.level}{the agreement level required. Default is 95\%. The proportion of data that should lie between the thresholds, for 95\% limits of agreement this should be 0.95.}

\item{alpha}{The alpha-level for confidence levels.}

\item{prop_bias}{Logical indicator (TRUE/FALSE) of whether proportional bias should be considered for the limits of agreement calculations.}

\item{log}{Calculate limits of agreement using log-transformed data.}
}
\value{
Returns single simple_agree class object with the results of the agreement analysis.

\describe{
  \item{\code{"loa"}}{A.}
}
}
\description{
A function for calculating for limits of agreement.
}
\section{References}{

Zou, G. Y. (2013). Confidence interval estimation for the Bland–Altman limits of agreement with multiple observations per individual. Statistical methods in medical research, 22(6), 630-642.
}

\examples{
data('reps')

}
