% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreetest.h.R
\name{agreetest}
\alias{agreetest}
\title{Simple Test of Agreement}
\usage{
agreetest(
  data,
  method1,
  method2,
  ciWidth = 0.95,
  agreeWidth = 0.95,
  testValue = 2
)
}
\arguments{
\item{data}{.}

\item{method1}{.}

\item{method2}{.}

\item{ciWidth}{a number between .50 and .999 (default: .95), the width of
confidence intervals}

\item{agreeWidth}{a number between .50 and .999 (default: .95), the width
of agreement limits}

\item{testValue}{a number specifying the}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Simple Test of Agreement
}
