% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reps}
\alias{reps}
\alias{ba1986}
\title{reps}
\format{
A data frame with 20 rows with 3 variables
\describe{
\item{id}{Subject identifier}
\item{x}{X measurement}
\item{y}{Y measurement}

}

A data frame with 17 rows with 5 variables
\describe{
\item{id}{Subject identifier}
\item{wright1}{PERF measurement #1 using Wright device}
\item{wright2}{PERF measurement #2 using Wright device}
\item{mini1}{PERF measurement #1 using Mini device}
\item{mini2}{PERF measurement #2 using Mini device}
}
}
\usage{
reps

ba1986
}
\description{
A fake data set of a agreement study where both measures have replicates.

The data set published in the original Bland & Altman paper on agreement.
}
\section{References}{


Bland, J. M., & Altman, D. (1986). Statistical methods for assessing agreement between two methods of clinical measurement. The Lancet, 327(8476), 307-310.
}

\keyword{datasets}
