% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.loa_mermod.R
\name{loa_mermod-methods}
\alias{loa_mermod-methods}
\alias{print.loa_mermod}
\alias{plot.loa_mermod}
\alias{check.loa_mermod}
\title{Methods for loa_mermod objects}
\usage{
\method{print}{loa_mermod}(x, ...)

\method{plot}{loa_mermod}(
  x,
  x_label = "Average of Both Methods",
  y_label = "Difference Between Methods",
  geom = "geom_point",
  smooth_method = NULL,
  smooth_se = TRUE,
  ...
)

\method{check}{loa_mermod}(x)
}
\arguments{
\item{x}{object of class \code{loa_mermod}.}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{loa_mixed}}.}

\item{x_label}{Label for x-axis.}

\item{y_label}{Label for y-axis.}

\item{geom}{String naming the type of geometry to display the data points. Default is "geom_point". Other options include: "geom_bin2d", "geom_density_2d", "geom_density_2d_filled", and "stat_density_2d".}

\item{smooth_method}{Smoothing method (function) to use, accepts either NULL or a character vector, e.g. "lm", "glm", "gam", "loess" or a function. Default is NULL, which will not include a trend line.}

\item{smooth_se}{Display confidence interval around smooth?}
}
\value{
\describe{
\item{\code{print}}{Prints short summary of the Limits of Agreement}
\item{\code{plot}}{Returns a plot of the limits of agreement}
}
}
\description{
Methods defined for objects returned from the loa_lme.
}
