% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loa_mixed.R
\name{loa_mixed}
\alias{loa_mixed}
\title{Mixed Effects Limits of Agreement}
\usage{
loa_mixed(
  diff,
  condition,
  id,
  data,
  conf.level = 0.95,
  agree.level = 0.95,
  replicates = 500,
  type = "bca"
)
}
\arguments{
\item{diff}{column name of the data frame that includes the continuous measurement of interest.}

\item{condition}{column name indicating different conditions subjects were tested under.}

\item{id}{column name indicating the subject/participant identifier}
}
\value{
Returns single list with the results of the agreement analysis.

\describe{
  \item{\code{"shieh_test"}}{The TOST hypothesis test as described by Shieh.}

}
}
\description{
Mixed Effects Limits of Agreement
}
\section{References}{

Parker, R. A., Weir, C. J., Rubio, N., Rabinovich, R., Pinnock, H., Hanley, J., McLoughan, L., Drost, E.M., Mantoani, L.C., MacNee, W., & McKinstry, B. (2016). Application of mixed effects limits of agreement in the presence of multiple sources of variability: exemplar from the comparison of several devices to measure respiratory rate in COPD patients. Plos one, 11(12), e0168321. <https://doi.org/10.1371/journal.pone.0168321>
}

\examples{
#to be added

}
