% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.simple_eiv.R
\name{simple_eiv-methods}
\alias{simple_eiv-methods}
\alias{print.simple_eiv}
\alias{plot.simple_eiv}
\alias{check.simple_eiv}
\title{Methods for simple_eiv objects}
\usage{
\method{print}{simple_eiv}(x, ...)

\method{plot}{simple_eiv}(x, x_name = "x", y_name = "y", ...)

\method{check}{simple_eiv}(x)
}
\arguments{
\item{x}{object of class \code{simple_eiv} from the dem_reg function.}

\item{...}{further arguments passed through, see description of return value.
for details.
\code{\link{agree_test}}.}

\item{x_name}{Name/label for x values (first measurement)}

\item{y_name}{Name/label for y values (second measurement)}
}
\value{
\describe{
\item{\code{print}}{Prints short summary of the error-in-variables (e.g., Deming) regression model.}
\item{\code{plot}}{Returns a plot of the deming regression line, the line-of-identity, and the raw data.}
\item{\code{check}}{Returns plots of the optimized residuals.}
}
}
\description{
Methods defined for objects returned from the error-in-variables models (e.g., dem_reg).
}
